aggregate-hospital/requirements.txt
kafka-python==2.0.2
flask==3.1.0